Reorganize benchmark.py into folder
