+++ cli/trainloop_cli/commands/benchmark/__init__.py
@@ -1,0 +1,5 @@
+"""TrainLoop CLI - benchmark module."""
+
+from .command import benchmark_command
+
+__all__ = ["benchmark_command"]
